requirements.txt
absl-py==0.14.1
appdirs==1.4.4
argon2-cffi==21.1.0
astroid==2.8.2
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Automat==20.2.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bcrypt==3.2.0
black==21.10b0
bleach==4.1.0
brotlipy==0.7.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.7
clang==5.0
click==8.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.8
cssselect==1.1.0
cycler==0.10.0
dataclasses==0.8
debugpy==1.4.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
filelock==3.3.2
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h5py==3.1.0
huggingface-hub==0.0.17
hyperlink==21.0.0
idna==3.1
importlib-metadata==4.8.1
incremental==21.3.0
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
isort==5.9.3
itemadapter==0.4.0
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.1.0
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-service==2.4.0
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.2
oauthlib==3.1.1
olefile==0.46
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parsel==1.6.0
parso==0.8.2
pathspec==0.9.0
patsy==0.5.2
pickleshare==0.7.5
Pillow==8.3.2
pip==21.3
platformdirs==2.3.0
plotly==5.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
PyDispatcher==2.0.5
pydot==1.4.2
Pygments==2.10.0
pylint==2.11.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pywin32==301
pywinpty==0.5.7
PyYAML==6.0
pyzmq==22.3.0
queuelib==1.6.2
regex==2021.10.23
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.43
scikit-learn==1.0
scipy==1.6.2
Scrapy==2.4.1
seaborn==0.11.2
Send2Trash==1.8.0
service-identity==18.1.0
setuptools==58.0.4
six==1.16.0
statsmodels==0.13.0
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.5.0
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
tomli==1.2.2
torch==1.9.1
torchaudio==0.9.1
torchvision==0.2.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
transformers==4.11.3
Twisted==21.7.0
twisted-iocpsupport==1.0.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wheel==0.37.0
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
zipp==3.6.0
zope.interface==5.4.0
